Fix broken tests
This is b/c of a faulty assumption that anonymous mmap's all map to the
same physical memory. D'oh.

Looks like the only way to have in-memory file is via ram disk. I'm on
mac, and I don't really want to mount one just for unit testing. So this
can be future work.