Optimize: Reader no longer clones Arc
Previously, `cargo bench` would report very slow throughput when
increasing thread count. Upon profiling, it turned out a lot of time was
wasted cloning an Arc over the Mmap.

Given increasing/decreasing ref count (atomically) is the bottleneck, I
replaced the clone with a borrow instead. This can only work in the
Reader case, as in the Writer, the Arc can be replaced completely due to
file growth.

An icky side effect is that we need to refactor the entire code base
with a ton of more generics. This'll probably need a cleanup afterwards.