Cleanup: remove unnecessary WriterPage Arc clone
This is a similar change to what was done to ReaderPage when
removing *its* Arc clone.

This wasn't really an optimization b/c the writer doesn't suffer Arc
contention; there can only at most 1 writer at a time. But, I went with
this change anyways for consistency with what was done with ReaderPage.

Again, this required a massive addition of 's annotations :(